requirements.txt
gradio
google-generativeai
murf
speechrecognition
pydub
requests
ffmpeg-python
